requirements.txt
streamlitnumpypandasPillowmatplotlibopencv-python-headlessscikit-learntorchtorchvisiontorchaudio